extract meal assist
