+++ lib/basal-set-temp.js
@@ -8,7 +8,8 @@ var setTempBasal = function (rate, duration, profile, rT, currenttemp) {
         rate = maxSafeBasal; 
     }
     
-    if (typeof(currenttemp) !== 'undefined' && typeof(currenttemp.duration) !== 'undefined' && typeof(currenttemp.rate) !== 'undefined' && currenttemp.duration > 20 && rate < currenttemp.rate + 0.1 && rate > currenttemp.rate - 0.1) {
+    if (typeof(currenttemp) !== 'undefined' && typeof(currenttemp.duration) !== 'undefined' && typeof(currenttemp.rate) !== 'undefined' 
+        && currenttemp.duration > 20 && rate < currenttemp.rate + 0.1 && rate > currenttemp.rate - 0.1) {
         rT.reason += ", but "+currenttemp.duration+"m left and " + currenttemp.rate + " ~ req " + rate + "U/hr: no action required";
         return rT;
     }
