+++ lib/determine-basal/determine-basal.js
@@ -12,7 +12,7 @@
   OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
   THE SOFTWARE.
 */
-var determine_basal = function determine_basal(glucose_status, currenttemp, iob_data, profile, autosens_data, meal_data, setTempBasal) {
+var determine_basal = function determine_basal(glucose_status, currenttemp, iob_data, profile, autosens_data, meal_data, mealAssistFn, setTempBasal) {
     var rT = { //short for requestedTemp
     };
 
@@ -48,6 +48,7 @@ var determine_basal = function determine_basal(glucose_status, currenttemp, iob_
     } else {
         if (typeof profile.min_bg !== 'undefined' && typeof profile.max_bg !== 'undefined') {
             target_bg = (profile.min_bg + profile.max_bg) / 2;
+            profile.target_bg = target_bg;
         } else {
             rT.error ='Error: could not determine target_bg';
             return rT;
@@ -131,54 +132,13 @@ var determine_basal = function determine_basal(glucose_status, currenttemp, iob_
     if (iob_data.basaliob) { basaliob = iob_data.basaliob; }
     else { basaliob = iob_data.iob - iob_data.bolussnooze; }
 
-    // net amount of basal insulin delivered over the last DIA hours
-    var hightempinsulin = iob_data.hightempinsulin;
-
-    var wtfAssist=0;
-    var mealAssist=0;
-    var mealAssistPct = 0;
-    //var wtfAssistPct = 0;
-    // if BG is high (more than DIA hours of basal above max_bg, i.e. above about 220mg/dL) and rising, wtf-assist
-    var high = profile.max_bg + ( basal * (profile.dia) * sens );
-    if ( bg > high && minDelta > Math.max(0,bgi) ) {
-        wtfAssist=1;
-    }
-    // minDelta is > 12 and devation is > 50 wtf-assist and meal-assist
-    var wtfDeviation=50;
-    var wtfDelta=12;
-    if ( deviation > wtfDeviation && minDelta > wtfDelta ) {
-        wtfAssist=1;
-        mealAssist=1;
-    } else {
-        // phase in mealAssist, as a fraction
-        mealAssist = Math.max(0, Math.round( Math.min(deviation/wtfDeviation,minDelta/wtfDelta)*100)/100 );
-    }
-    var remainingMealBolus = Math.round( (1.1 * meal_data.carbs/profile.carb_ratio - ( meal_data.boluses + Math.max(0,hightempinsulin) ) )*10)/10;
-        // if minDelta is >3 and >BGI, and there are uncovered carbs, meal-assist
-    if ( minDelta > Math.max(3, bgi) && meal_data.carbs > 0 && remainingMealBolus > 0 ) {
-        mealAssist=1;
-    }
-    // when rising with carbs or rising fast for no good reason, meal-assist (ignore bolus IOB)
-    if (mealAssist > 0) {
-        // ignore all covered IOB, and just set eventualBG to the current bg
-        mAeventualBG = Math.max(bg,eventualBG) + deviation;
-        eventualBG = Math.round(mealAssist*mAeventualBG + (1-mealAssist)*eventualBG);
-        rT.eventualBG = eventualBG;
-        //console.error("eventualBG: "+eventualBG+", mAeventualBG: "+mAeventualBG+", rT.eventualBG: "+rT.eventualBG);
-    }
-    // lower target for meal-assist or wtf-assist (high and rising)
-    wtfAssist = Math.round( Math.max(wtfAssist, mealAssist) *100)/100;
-    if (wtfAssist > 0) {
-        min_bg = wtfAssist*80 + (1-wtfAssist)*min_bg;
-        target_bg = (min_bg + profile.max_bg) / 2;
-        expectedDelta = Math.round(( bgi + ( target_bg - eventualBG ) / ( profile.dia * 60 / 5 ) )*10)/10;
-        mealAssistPct = Math.round(mealAssist*100);
-        wtfAssistPct = Math.round(wtfAssist*100);
-        rT.mealAssist = "On: "+mealAssistPct+"%, "+wtfAssistPct+"%, Carbs: " + meal_data.carbs + " Boluses: " + meal_data.boluses + " Target: " + Math.round(target_bg) + " Deviation: " + deviation + " BGI: " + bgi;
-    } else {
-        rT.mealAssist = "Off: Carbs: " + meal_data.carbs + " Boluses: " + meal_data.boluses + " Target: " + Math.round(target_bg) + " Deviation: " + deviation + " BGI: " + bgi;
-    }
-
+    var mealAssistResult= mealAssistFn(meal_data, profile, iob_data, basal, sens, bg, deviation, minDelta, bgi, eventualBG, rT);
+    var remainingMealBolus = mealAssistResult.remainingMealBolus;
+    var mealAssist = mealAssistResult.mealAssist;
+    min_bg = mealAssistResult.min_bg;
+    if(!(typeof mealAssistResult.expectedDelta === 'undefined')) expectedDelta = mealAssistResult.expectedDelta;
+    
+    
     rT.reason="";
     if (bg < threshold) { // low glucose suspend mode: BG is < ~80
         rT.reason += "BG " + bg + "<" + threshold;
@@ -361,11 +321,10 @@ var determine_basal = function determine_basal(glucose_status, currenttemp, iob_
         // rate required to deliver insulinReq more insulin over 30m:
         var rate = basal + (2 * insulinReq);
         rate = Math.round( rate * 1000 ) / 1000;
-
             var maxSafeBasal = Math.min(profile.max_basal, 3 * profile.max_daily_basal, 4 * basal);
         if (rate > maxSafeBasal) {
             rT.reason += "adj. req. rate:"+rate.toFixed(1) +" to maxSafeBasal:"+maxSafeBasal.toFixed(1)+", ";
-            rate = maxSafeBasal.toFixed(1);
+            rate = maxSafeBasal;
         }
         
         var insulinScheduled = currenttemp.duration * (currenttemp.rate - basal) / 60;
